apps/engine/pyproject.toml
[project]
name = "n1hub-engine"
version = "0.1.0"
description = "N1Hub v0.1 DeepMine engine"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-ulid>=1.1.0",
  "orjson>=3.10.0",
  "asyncpg>=0.29.0",
  "redis>=5.0.0",
  "anthropic>=0.18.0",
  "openai>=1.12.0",
  "sentence-transformers>=2.2.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.1.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
